Migrate error types to effect modules
