+++ src/Strategy/Carthage.hs
@@ -18,7 +18,6 @@ import Text.Megaparsec
 import Text.Megaparsec.Char
 import qualified Text.Megaparsec.Char.Lexer as L
 
-import Diagnostics
 import Discovery.Walk
 import DepTypes
 import Effect.Grapher
