+++ src/Strategy/Node/NpmList.hs
@@ -7,7 +7,6 @@ import Prologue
 
 import Control.Carrier.Error.Either
 import qualified Data.Map.Strict as M
-import Diagnostics
 import DepTypes
 import Discovery.Walk
 import Effect.Exec
