+++ src/Strategy/Node/PackageJson.hs
@@ -9,7 +9,6 @@ import Prologue
 
 import Control.Carrier.Error.Either
 import qualified Data.Map.Strict as M
-import Diagnostics
 import DepTypes
 import Discovery.Walk
 import Effect.LabeledGrapher
