+++ src/Strategy/NuGet/Nuspec.hs
@@ -14,7 +14,6 @@ import Control.Carrier.Error.Either
 import qualified Data.Map.Strict as M
 import qualified Data.List as L
 
-import Diagnostics
 import DepTypes
 import Discovery.Walk
 import Effect.ReadFS
