+++ src/Strategy/NuGet/PackagesConfig.hs
@@ -11,7 +11,6 @@ import Prologue
 
 import Control.Carrier.Error.Either
 import qualified Data.Map.Strict as M
-import Diagnostics
 import DepTypes
 import Discovery.Walk
 import Effect.ReadFS
