+++ src/Strategy/NuGet/ProjectAssetsJson.hs
@@ -13,7 +13,6 @@ import qualified Data.Map.Strict as M
 import qualified Data.Text as T
 import Data.Maybe
 
-import Diagnostics
 import DepTypes
 import Discovery.Walk
 import Effect.ReadFS
