+++ src/Strategy/Python/ReqTxt.hs
@@ -11,7 +11,6 @@ import Control.Carrier.Error.Either
 import Text.Megaparsec
 import Text.Megaparsec.Char
 
-import Diagnostics
 import Discovery.Walk
 import Effect.ReadFS
 import Strategy.Python.Util
